Further related refactors for #149
